dev-requirements.txt
# Dev / test specific minimal pins
requests==2.32.5
prometheus_client==0.23.1
pytest-asyncio==1.2.0
python-json-logger==3.3.0
PyJWT==2.10.1

# Dev tooling
mypy==1.6.0
ruff==0.13.1
flake8==6.0.0
pytest>=8.2,<9

# Typing stubs to improve mypy coverage in developer envs
types-requests
types-PyYAML
types-psutil
types-redis
types-PyJWT
types-passlib
types-sqlalchemy
types-psycopg2

# Test client compatibility pins
httpx==0.24.1
httpcore==0.17.0
# Keep starlette compatible with fastapi pinned in requirements.txt
# FastAPI 0.104.1 requires starlette<0.28.0,>=0.27.0
starlette==0.27.0
fastapi==0.104.1
itsdangerous

# Runtime libraries required by application imports during test collection
# Use a wheel that supports newer Python versions (incl. 3.13)
PyYAML==6.0.2
email-validator==2.1.1
structlog==24.2.0
redis==5.0.1
passlib==1.7.4
pydantic-settings==2.11.0

# Optional dependencies used by performance and async tests/tools
# Skip for Python >= 3.13 to avoid missing wheels/build toolchain on Windows
aiohttp==3.8.4; python_version < "3.13"
asyncpg==0.27.0; python_version < "3.13"
python-memcached==1.59
psutil==5.9.5
jinja2==3.1.2
